Renaming ContextSubscriber to ServiceSubscriber and ContextProvider to ServiceProvider. Moving getServices() and service() to Context.php
