+++ src/Context/SiteContext.php
@@ -3,8 +3,7 @@
 namespace Aegir\Provision\Context;
 
 use Aegir\Provision\Application;
-use Aegir\Provision\Context;
-use Aegir\Provision\ContextSubscriber;
+use Aegir\Provision\ServiceSubscriber;
 use Aegir\Provision\Provision;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -15,7 +14,7 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
  *
  * @see \Provision_Context_site
  */
-class SiteContext extends ContextSubscriber implements ConfigurationInterface
+class SiteContext extends ServiceSubscriber implements ConfigurationInterface
 {
     /**
      * @var string
